refactor: creation core dir and refactoring the controller
